LL3/requirements.txt
crewai
flask
python-dotenv
PyYAML

